Initial commit
Bit broken.
The SpatialQueryArrayLegacy class has it's origins from FastStar
and was made by me with the initial concept coming from my mother.
The "new" SpatialQueryArray was made by Ret-Mode, see
https://gist.github.com/Ret-Mode/98cfe91a655e8f496902071a372e4f6c

That being said, the new implementation is slower for 1nn calls,
is hardcoded for 250k star galaxies (completely ruining the purpose
of this library), and has a few bugs according to the implementor.

There is still a bit to work on, so this repo isn't yet connected
to our ""CI/CD"" pipeline.