Publish as an artifact that can actually be consumed by nightly-paperpusher
oop. For reference, nightly-paperpusher is our backend for the maven repository
and it doesn't handle -SNAPSHOT versions very well. I should change that honestly,
but maven has slightly different artifact uploading procedures for these artifacts,
so it wouldn't be a completely easy change. And the workaround is easy enough for
now.